more component renaming
