+++ tests/unit/components/Home/CarouselContent.spec.js
@@ -1,5 +1,5 @@
 import { shallowMount } from "@vue/test-utils";
-import TabContent from "@/components/Home/BlockTabs/TabContent"
+import CarouselContent from "@/components/Home/CarouselContent"
 import Vuetify from "vuetify"
 
 const vuetify = new Vuetify();
@@ -8,14 +8,14 @@ describe("TabContent.vue", function(){
     let wrapper;
 
     beforeEach(() => {
-        wrapper = shallowMount(TabContent, {
+        wrapper = shallowMount(CarouselContent, {
             vuetify,
             stubs: ['router-link']
         })
     });
 
     it("can be instantiated", () => {
-        expect(wrapper.name()).toMatch("TabContent");
+        expect(wrapper.name()).toMatch("CarouselContent");
     });
 
 });
